Added docker-compose to run examples compilation
