+++ docker-compose.yml
@@ -1,0 +1,17 @@
+version: '3'
+
+services:
+      examples:
+            image: "fabiosoaza/snes-dev-tools:0.1.0-SNAPSHOT" 
+            command: /bin/bash -c "cd /tmp/hello-world/ && ./build.sh HelloWorld && cd /tmp/png2snes-hello-world/ && make && cd /tmp/walker-example/ && ./build_game.sh"
+            volumes: 
+                  - ./examples/hello-world:/tmp/hello-world
+                  - ./examples/png2snes/hello-world:/tmp/png2snes-hello-world
+                  - ./examples/Neviksti/walker-example:/tmp/walker-example
+
+networks:
+  default:
+    ipam:
+      driver: default
+      config:
+        - subnet: 10.20.0.1/16
